refactor: move the mainscreen composeable to a new file
